+++ lib/alf.rb
@@ -25,6 +25,7 @@ module Alf
   require 'alf/lang'
   require 'alf/relation'
   require 'alf/lispy'
+  require 'alf/ext'
 
   #
   # Builds and returns a lispy engine on a specific environment.